Allow Observable#get in expressions
This CL reverts a change from 3.6 which disabled calling Observable#get
in expressions. Unfortunately, two way binding generates similar expressions
just to be executed directly so disabling these might break nested observable
bindings.

IDE still shows an error and it should since this is an error if user does it
but generated code is fine. We could potentially fix this by allowing just the
generated code to do it but it would be a much larger change that i'm not super
comfortable with.

Bug: 147249308
Test: ObservableGetDetectionTest TwoWayBindingAdapterTest
Change-Id: I592bb6404992df6ab3ada34851c59028a6bae994
(cherry picked from commit be15847d80211e2388bd11faa86f04bf2f638523)