Set --max-workers=1 to reduce DB tests' flakiness on Windows
Limit the number of Gradle workers to reduce the chance that the Gradle
build process hangs.

We do this on Windows only as the tests do not flake as frequently on
other OS's.

Also, we will be collecting the thread dump (Issue 179482828), so that
the next time the tests hang, we will have some data to analyze.

Bug: 177936851
Test: Existing tests
Change-Id: Ic75c6797efea25243fb98db148920eda7aa17d3f