Avoid StackOverflow in code generation by dispatching a real error
When a callback expression returns void but it is expected to return a
value, data binding goes into an infinite loop because it tries to find
a value but cannot because void cannot be a assigned to a variable.

The right fix here would be to check for assignability and throw a
better error message but it might break some existing apps because i
don't trust our assignability checks. Instead, this CL covers the void
return type case to fix this one bug without taking a bigger risk.

Fixes: 123260053
Test: SimpleCompilationTest.testCallbackReturnTypeMismatch

Change-Id: Ie9b4d03e0e2d818a0613f4cdefc470e1f7221b27