Sort files in directories to ensure deterministic order
java.io.File.listFiles() does not guarantee consistent order of the
returned files. The same is true for
org.apache.commons.io.FileUtils.listFiles() which data binding uses.

This commit adds util methods to sort the returned files to ensure
deterministic order, which is important for deterministic and faster
builds.

Bug: 187340555
Test: Existing tests (new test will be added in the next change)
Change-Id: I10d0053a67f7784653b5e9745a848d62820f374a