Add support for tools:viewBindingType
This new attribute allows developer to change the type of the view for
ViewBinding.
This is a critical use case for large screen layouts where different
configurations have different sub types for a common base class and this
configuration parameter lets user handle those cases better by
specifying the parent type.

These values are not validated aside from a simple check across
configurations because view binding does not know about classpaths.

This attribute is not supported for `<include>` tags since they refer to
a generated class when the included layout is a binding. We could
potentially support include cases when the layout is not a binding
layout as a followup CL (if there is demand).

Bug: 154531811
Test: ViewBindingTypeTest, ViewBinderGenerateJavaTest

Change-Id: I22456a1b4e133a24d51f155aae2ad02149e02a74