Fix sdk file loading
The api-versions file was moved. It was added to the new location in
platforms in API 26 was removed in platform-tools 31+, in order that it
is versioned alongside with the platform.

Previously, the SDK location would be an argument to the data binding
compiler, with this and the corresponding change in the Android Gradle
Plugin, the the api-versions.xml for the current api level is directly
injected, which means that it can be treated as an input from Gradle's
perspective, and that the build won't be broken again by SDK component
updates.

AGP does not fall back to looking for the api file in platform-tools,
instead the existing bundled version in

Bug: 196847367
Test: SdkVersionTest
Change-Id: I849289c7d41672ca5caf8c4250e04e35b4ce39ae