Unroll deprecated lifecycle-extensions dependency
This CL updates data binding runtime dependencies to not
depend on deprecated lifecycle-extensions.

It was rather a wrapper to get consistent dependencies for
other lifecycle artifacts but since that package is deprecated,
it is useless.

I thought about updating dependencies to more recent versions
but that might affect apps. Instead, I just unrolled to remove
the deprecated dependency.

Fixes: 197553072
Test: existing tests
Change-Id: I6c3482ee21d6c0f650c46e4e6fe92d689dfa2601