Non-transitive R classes: Allow empty dependencies
It is possible to have a library with no dependencies, and in the
process of serialization of compiler arguments the difference between
empty string (serializing an empty list) and null is lost.

Just relax the assertion here, a reference to a resource that doesn't
exist will still be caught by the compiler.

(cherry picked from commit 413cb696aaad255d0ac529946ea26c87cbe00e82)

Bug: 216667498
Test: Added unit test
Change-Id: Iaeb8494de85e9356ea68134804f9e92f2ba754c8