Compile data binding runtime using Gradle AGP
As this makes the path to compiling the runtime somewhat more
monolithic, this might start showing up in the critical path more often.

If that becomes a problem, we can investigate compiling the data binding
runtime for tests using a fixed version of AGP, and then having a
separate test that validates that the produced library is the same as is
produced by running the current AGP at head, with the option to
temporarily switch back if there are changes.

Bug: 143864616
Test: Migrating tests to use our production build
Change-Id: Ic5d35a8a2a5f9bd6bd37fb98084dc61534afd432