Remove DataBindingCompilationTestCase base class
All tests that use DataBindingCompilationTestCase are currently disabled
for Windows due to timeouts. Those timeouts were caused by deadlocks
happening during initialization in the base AndroidGradleTestCase.

That base class is unnecessary, since all the derived classes specify a
runner and the logic that's used can be provided by
AndroidGradleProjectRule instead. Switching to the rule removes any
timeouts on Windows.

This test failed when run from HEAD on Windows ~4-5% of the time. With
this change, it's passed for 500 runs:
https://fusion2.corp.google.com/invocations/0fd0691d-3658-45c3-bb2a-fb91865af04c

Test: n/a (test code)
Fixes: 177936851
Change-Id: I46422a074733480e381e9abd36d0cf41898235fc