Disable data binding TooManyLayoutsTest on Windows
TooManyLayoutsTest is a stress test. It takes a long time to run and
also has high variance in running time.

Disabling it on Windows to reduce build latency (b/301266550) and avoid
potential timeouts (b/300720819).

We probably don't need to re-enable this test later as having it run on
other platforms is probably enough.

Test: Disabling TooManyLayoutsTest on Windows
Bug: 301266550, 300720819
Change-Id: Ie37c454748dd701ce82e625f8022dd5872883a56